refactor: 앱 초기화와 화면 구성을 역할별로 분리
