+++ src/AppRouter.tsx
@@ -1,10 +1,66 @@
+import type { RouteObject } from 'react-router-dom';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-import { appRoutes } from './routes';
+import AuthGuard from '@components/common/Auth/AuthGuard';
+import ChatPage from '@pages/ChatPage/ChatPage';
+import DocumentPage from '@pages/DocumentPage/DocumentPage';
+import EntryPage from '@pages/EntryPage/EntryPage';
+import FeedPage from '@pages/FeedPage/FeedPage';
+import InvitePage from '@pages/InvitePage/InvitePage';
+import LandingPage from '@pages/LandingPage/LandingPage';
+import MyPage from '@pages/MyPage/MyPage';
+import NotFoundPage from '@pages/NotFoundPage/NotFoundPage';
+import PostPage from '@pages/PostPage/PostPage';
+import RedirectPage from '@pages/RedirectPage/RedirectPage';
+import SettingPage from '@pages/SettingPage/SettingPage';
+import SignInPage from '@pages/SignInPage/SignInPage';
+import SignUpPage from '@pages/SignUpPage/SignUpPage';
+import { PATH } from '@constants/path';
+import NavigationLayout from './layouts/NavigationLayout';
+import PageLayout from './layouts/PageLayout';
+import RootLayout from './layouts/RootLayout';
 
-const AppRouter = () => {
-	const router = createBrowserRouter(appRoutes);
+const appRoutes: RouteObject[] = [
+	{
+		path: PATH.ROOT,
+		element: <RootLayout />,
+		errorElement: <NotFoundPage />,
+		children: [
+			{
+				element: <PageLayout />,
+				children: [
+					{ path: '', element: <LandingPage /> },
+					{ path: PATH.SIGNIN, element: <SignInPage /> },
+					{ path: PATH.SIGNUP, element: <SignUpPage /> },
+					{ path: `${PATH.REDIRECT}/:provider`, element: <RedirectPage /> },
+					{ path: PATH.INVITE, element: <InvitePage /> },
+				],
+			},
+			{
+				element: <AuthGuard />,
+				children: [
+					{
+						element: <PageLayout />,
+						children: [{ path: PATH.ENTRY, element: <EntryPage /> }],
+					},
+					{
+						element: <NavigationLayout />,
+						children: [
+							{ path: PATH.SETTING, element: <SettingPage /> },
+							{ path: PATH.MYPAGE, element: <MyPage /> },
+							{ path: PATH.FEED, element: <FeedPage /> },
+							{ path: PATH.POST, element: <PostPage /> },
+							{ path: PATH.CHAT, element: <ChatPage /> },
+							{ path: PATH.DOCUMENT, element: <DocumentPage /> },
+						],
+					},
+				],
+			},
+		],
+	},
+];
 
-	return <RouterProvider router={router} />;
-};
+const router = createBrowserRouter(appRoutes);
+
+const AppRouter = () => <RouterProvider router={router} />;
 
 export default AppRouter;
