+++ src/main.tsx
@@ -1,13 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { queryClient } from '@hooks/queries/common/queryClient';
-import { QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { ThemeProvider } from 'styled-components';
 import { worker } from '@mocks/browser';
-import GlobalStyle from '@styles/GlobalStyle.ts';
-import theme from '@styles/theme.ts';
-import AppRouter from './AppRouter';
+import App from './App';
 import './index.css';
 
 if (process.env.NODE_ENV === 'development') {
@@ -16,12 +10,6 @@ if (process.env.NODE_ENV === 'development') {
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
 	<React.StrictMode>
-		<ThemeProvider theme={theme}>
-			<QueryClientProvider client={queryClient}>
-				<GlobalStyle />
-				<AppRouter />
-				<ReactQueryDevtools />
-			</QueryClientProvider>
-		</ThemeProvider>
+		<App />
 	</React.StrictMode>
 );
